refactor: extract monaco configs
