+++ src/views/editor/index.vue
@@ -11,7 +11,6 @@
 <script setup lang="ts">
 import { open } from '@tauri-apps/api/shell';
 import { listen } from '@tauri-apps/api/event';
-import * as monaco from 'monaco-editor';
 import { storeToRefs } from 'pinia';
 import { onMounted, onUnmounted, ref, watch } from 'vue';
 import { useMessage } from 'naive-ui';
@@ -23,10 +22,10 @@ import {
   EngineType,
   getActionApiDoc,
   SearchAction,
-  searchTokensProvider,
 } from '../../common';
 import { useAppStore, useChatStore, useConnectionStore, useSourceFileStore } from '../../store';
 import { useLang } from '../../lang';
+import { monaco } from '../../common/monaco';
 
 type Editor = ReturnType<typeof monaco.editor.create>;
 
@@ -46,62 +45,6 @@ const { themeType } = storeToRefs(appStore);
 
 const chatStore = useChatStore();
 const { insertBoard } = storeToRefs(chatStore);
-/**
- * refer https://github.com/wobsoriano/codeplayground
- * https://github.com/wobsoriano/codeplayground/blob/master/src/components/MonacoEditor.vue
- */
-self.MonacoEnvironment = {
-  async getWorker(_, label) {
-    let worker;
-
-    switch (label) {
-      case 'json':
-        // @ts-ignore
-        worker = await import('monaco-editor/esm/vs/language/json/json.worker?worker');
-        break;
-      case 'css':
-      case 'scss':
-      case 'less':
-        // @ts-ignore
-        worker = await import('monaco-editor/esm/vs/language/css/css.worker?worker');
-        break;
-      case 'html':
-      case 'handlebars':
-      case 'razor':
-        // @ts-ignore
-        worker = await import('monaco-editor/esm/vs/language/html/html.worker?worker');
-        break;
-      case 'typescript':
-      case 'javascript':
-        // @ts-ignore
-        worker = await import('monaco-editor/esm/vs/language/typescript/ts.worker?worker');
-        break;
-      default:
-        // @ts-ignore
-        worker = await import('monaco-editor/esm/vs/editor/editor.worker?worker');
-    }
-
-    return new worker.default();
-  },
-};
-monaco.languages.typescript.typescriptDefaults.setEagerModelSync(true);
-
-monaco.languages.register({ id: 'search' });
-
-monaco.languages.setMonarchTokensProvider(
-  'search',
-  searchTokensProvider as monaco.languages.IMonarchLanguage,
-);
-monaco.languages.setLanguageConfiguration('search', {
-  autoClosingPairs: [
-    { open: '{', close: '}' },
-    { open: '[', close: ']' },
-    { open: '(', close: ')' },
-    { open: '"', close: '"' },
-    { open: "'", close: "'" },
-  ],
-});
-
 // https://github.com/tjx666/adobe-devtools/commit/8055d8415ed3ec5996880b3a4ee2db2413a71c61
 let displayEditor: Editor | null = null;
 let queryEditor: Editor | null = null;
